feat: expose reporter api
